+++ lib/frame_composer/ffmpeg_compositor_module.ex
@@ -1,0 +1,17 @@
+defmodule Membrane.VideoCompositor.FFMPEG do
+  @moduledoc """
+  This module implements video composition in ffmpeg.
+  """
+  @behaviour Membrane.VideoCompositor.FrameCompositor
+
+  @impl Membrane.VideoCompositor.FrameCompositor
+  def init(_caps) do
+    {:ok, %{}}
+  end
+
+  @impl Membrane.VideoCompositor.FrameCompositor
+  def merge_frames(frames, internal_state) do
+    merged_frames_binary = frames.first <> frames.second
+    {{:ok, merged_frames_binary}, internal_state}
+  end
+end
