+++ lib/frame_compositor.ex
@@ -1,0 +1,15 @@
+defmodule Membrane.VideoCompositor.FrameCompositor do
+  @moduledoc """
+  This module defines behaviour for different frame composition implementations.
+  """
+  alias Membrane.RawVideo
+
+  @callback init(caps :: RawVideo) :: {:ok, state :: any()}
+  @callback merge_frames(
+              frames :: %{
+                first: binary(),
+                second: binary()
+              },
+              internal_state :: any()
+            ) :: {{:ok, merged_frames :: binary()}, state :: any()}
+end
