+++ mix.exs
@@ -33,7 +33,7 @@ defmodule Membrane.VideoCompositor.Mixfile do
   end
 
   defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_env), do: ["lib"]
+  defp elixirc_paths(_env), do: ["lib", "test/support"]
 
   defp deps do
     [
