+++ test/composing_test.exs
@@ -1,51 +1,101 @@
-ExUnit.start()
-
 defmodule Membrane.VideoCompositor.ComposingTest do
   use ExUnit.Case
 
   import Membrane.Testing.Assertions
 
-  alias Membrane.Testing.Pipeline, as: TestingPipeline
   alias Membrane.RawVideo
+  alias Membrane.Testing.Pipeline, as: TestingPipeline
 
-  test "Checks composition and raw video pipeline" do
-    in_path = "./test/fixtures/input_10s_720p_1fps.raw"
-    out_path = "./test/fixtures/output_10s_1280x1440_1fps.raw"
-    composed_video_path = "./test/fixtures/composed_video_10s_1280x1440_1fps.raw"
+  describe "Checks composition and raw video pipeline on " do
+    @describetag :tmp_dir
 
-    video_caps = %RawVideo{
-      width: 1280,
-      height: 720,
-      framerate: {1, 1},
-      pixel_format: :I420,
-      aligned: nil
-    }
+    test "on 3s 720p 1fps raw video", ctx do
+      input_paths = %{
+        first_raw_video_path: "./test/fixtures/short_videos/input_3s_720p_1fps.raw",
+        second_raw_video_path: "./test/fixtures/short_videos/input_3s_720p_1fps.raw"
+      }
 
-    implementation = :nx
-    duration = 10
+      output_path = Path.join(ctx.tmp_dir, "output_3s_1280x1440_1fps.raw")
+      composed_video_path = "./test/fixtures/short_videos/composed_video_3s_1280x1440_1fps.raw"
 
+      video_caps = %RawVideo{
+        width: 1280,
+        height: 720,
+        framerate: {1, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
+
+      implementation = :nx
+
+      test_raw_pipeline_and_composing(
+        input_paths,
+        output_path,
+        composed_video_path,
+        video_caps,
+        implementation
+      )
+    end
+
+    @tag long: true
+    test "10s 720p 1fps raw video", ctx do
+      input_paths = %{
+        first_raw_video_path: "./test/fixtures/long_videos/input_10s_720p_1fps.raw",
+        second_raw_video_path: "./test/fixtures/long_videos/input_10s_720p_1fps.raw"
+      }
+
+      output_path = Path.join(ctx.tmp_dir, "output_10s_1280x1440_1fps.raw")
+      composed_video_path = "./test/fixtures/long_videos/composed_video_10s_1280x1440_1fps.raw"
+
+      video_caps = %RawVideo{
+        width: 1280,
+        height: 720,
+        framerate: {1, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
+
+      implementation = :nx
+
+      test_raw_pipeline_and_composing(
+        input_paths,
+        output_path,
+        composed_video_path,
+        video_caps,
+        implementation
+      )
+    end
+  end
+
+  defp test_raw_pipeline_and_composing(
+         input_paths,
+         output_path,
+         composed_video_path,
+         video_caps,
+         implementation
+       ) do
     options = %{
       paths: %{
-        first_raw_video_path: in_path,
-        second_raw_video_path: in_path,
-        output_path: out_path
+        first_raw_video_path: input_paths.first_raw_video_path,
+        second_raw_video_path: input_paths.second_raw_video_path,
+        output_path: output_path
       },
       caps: video_caps,
       implementation: implementation
     }
 
     assert {:ok, pid} =
-             TestingPipeline.start_link(%TestingPipeline.Options{
+             TestingPipeline.start_link(
                module: Membrane.VideoCompositor.PipelineRaw,
                custom_args: options
-             })
+             )
 
     assert_pipeline_playback_changed(pid, _, :playing)
 
-    assert_end_of_stream(pid, :file_sink, :input, 100_000)
+    assert_end_of_stream(pid, :file_sink, :input, 1_000_000)
     TestingPipeline.terminate(pid, blocking?: true)
 
-    assert {:ok, out_video} = File.read(out_path)
+    assert {:ok, out_video} = File.read(output_path)
     assert {:ok, composed_video} = File.read(composed_video_path)
 
     assert out_video == composed_video
