+++ test/pipeline_integration_test.exs
@@ -1,84 +1,121 @@
-ExUnit.start()
-
 defmodule Membrane.VideoCompositor.PipelineTest do
   use ExUnit.Case
 
   import Membrane.Testing.Assertions
 
-  alias Membrane.Testing.Pipeline, as: TestingPipeline
   alias Membrane.RawVideo
+  alias Membrane.Testing.Pipeline, as: TestingPipeline
 
-  test "Checks h264 pipeline of 30s 720 videos" do
-    in_path = "./test/fixtures/input_30s_720p.h264"
-    out_path = "./test/fixtures/output_30s_1280x1440.h264"
+  describe "Checks h264 pipeline on " do
+    @describetag :tmp_dir
 
-    video_caps = %RawVideo{
-      width: 1280,
-      height: 720,
-      framerate: {30, 1},
-      pixel_format: :I420,
-      aligned: nil
-    }
+    test "2s 720p 30fps video", ctx do
+      input_paths = %{
+        first_h264_video_path: "./test/fixtures/short_videos/input_2s_720p.h264",
+        second_h264_video_path: "./test/fixtures/short_videos/input_2s_720p.h264"
+      }
 
-    implementation = :nx
-    duration = 30
+      output_path = Path.join(ctx.tmp_dir, "output_2s_1280x1440.h264")
 
-    options = %{
-      paths: %{
-        first_h264_video_path: in_path,
-        second_h264_video_path: in_path,
-        output_path: out_path
-      },
-      caps: video_caps,
-      implementation: implementation
-    }
+      video_caps = %RawVideo{
+        width: 1280,
+        height: 720,
+        framerate: {30, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
 
-    assert {:ok, pid} =
-             TestingPipeline.start_link(%TestingPipeline.Options{
-               module: Membrane.VideoCompositor.PipelineH264,
-               custom_args: options
-             })
+      implementation = :nx
 
-    assert_pipeline_playback_changed(pid, _, :playing)
+      test_h264_pipeline(input_paths, output_path, video_caps, implementation)
+    end
 
-    assert_end_of_stream(pid, :file_sink, :input, 100_000)
-    TestingPipeline.terminate(pid, blocking?: true)
+    test "1s 1080p 30fps video", ctx do
+      input_paths = %{
+        first_h264_video_path: "./test/fixtures/short_videos/input_1s_1080p.h264",
+        second_h264_video_path: "./test/fixtures/short_videos/input_1s_1080p.h264"
+      }
+
+      output_path = Path.join(ctx.tmp_dir, "output_1s_1920x2160.h264")
+
+      video_caps = %RawVideo{
+        width: 1920,
+        height: 1080,
+        framerate: {30, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
+
+      implementation = :nx
+
+      test_h264_pipeline(input_paths, output_path, video_caps, implementation)
+    end
+
+    @tag long: true
+    test "30s 720p 30fps video", ctx do
+      input_paths = %{
+        first_h264_video_path: "./test/fixtures/long_videos/input_30s_720p.h264",
+        second_h264_video_path: "./test/fixtures/long_videos/input_30s_720p.h264"
+      }
+
+      output_path = Path.join(ctx.tmp_dir, "output_30s_1280x1440.h264")
+
+      video_caps = %RawVideo{
+        width: 1280,
+        height: 720,
+        framerate: {30, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
+
+      implementation = :nx
+
+      test_h264_pipeline(input_paths, output_path, video_caps, implementation)
+    end
+
+    @tag long: true
+    test "60s 1080p 30fps video", ctx do
+      input_paths = %{
+        first_h264_video_path: "./test/fixtures/long_videos/input_60s_1080p.h264",
+        second_h264_video_path: "./test/fixtures/long_videos/input_60s_1080p.h264"
+      }
+
+      output_path = Path.join(ctx.tmp_dir, "output_60s_1920x2160.h264")
+
+      video_caps = %RawVideo{
+        width: 1920,
+        height: 1080,
+        framerate: {30, 1},
+        pixel_format: :I420,
+        aligned: nil
+      }
+
+      implementation = :nx
+
+      test_h264_pipeline(input_paths, output_path, video_caps, implementation)
+    end
   end
 
-  test "Checks h264 pipeline on 60s 1080p videos" do
-    in_path = "./test/fixtures/input_60s_1080p.h264"
-    out_path = "./test/fixtures/output_60s_1920x2160.h264"
-
-    video_caps = %RawVideo{
-      width: 1920,
-      height: 1080,
-      framerate: {30, 1},
-      pixel_format: :I420,
-      aligned: nil
-    }
-
-    implementation = :nx
-    duration = 30
-
+  defp test_h264_pipeline(input_paths, output_path, video_caps, implementation) do
     options = %{
       paths: %{
-        first_h264_video_path: in_path,
-        second_h264_video_path: in_path,
-        output_path: out_path
+        first_h264_video_path: input_paths.first_h264_video_path,
+        second_h264_video_path: input_paths.first_h264_video_path,
+        output_path: output_path
       },
       caps: video_caps,
       implementation: implementation
     }
 
     assert {:ok, pid} =
-             TestingPipeline.start_link(%TestingPipeline.Options{
+             TestingPipeline.start_link(
                module: Membrane.VideoCompositor.PipelineH264,
                custom_args: options
-             })
+             )
 
     assert_pipeline_playback_changed(pid, _, :playing)
 
-    assert_end_of_stream(pid, :file_sink, :input, 100_000)
+    assert_end_of_stream(pid, :file_sink, :input, 1_000_000)
     TestingPipeline.terminate(pid, blocking?: true)
   end
 end
