+++ test/test_helper.exs
@@ -1,1 +1,1 @@
-ExUnit.start(capture_log: true)
+ExUnit.start(capture_log: true, exclude: [long: true])
